+ Fixing the bamboo
git-svn-id: http://agilefant.googlecode.com/svn/branches/Monza_restructure@2396 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ test/fi/hut/soberit/agilefant/business/IterationBusinessTest.java
@@ -155,8 +155,8 @@ public class IterationBusinessTest {
         
         assertNotNull(actualMetrics);
         
-        assertEquals(latestEntry.getEffortLeftSum(), actualMetrics.getEffortLeft().getMinorUnits());
-        assertEquals(latestEntry.getOriginalEstimateSum(), actualMetrics.getOriginalEstimate().getMinorUnits());
+        assertEquals(latestEntry.getEffortLeftSum(), actualMetrics.getEffortLeft().getMinorUnits().longValue());
+        assertEquals(latestEntry.getOriginalEstimateSum(), actualMetrics.getOriginalEstimate().getMinorUnits().longValue());
         
 //        assertEquals(expectedStoryPoint, actualMetrics.getStoryPoints());
